requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
boto3>=1.35.0
rich>=13.0.0
pytest>=8.0.0
